+++ CHANGELOG.md
@@ -1,10 +1,20 @@
+
+
 # Changelog
 
 Lists all changes with user impact.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-
 ## [Unreleased]
+
+### Changed
+
+## [0.19.8]
+
+### Changed
+- Remove reactor in computing cluster state changes
+
+## [0.19.7]
 
 ### Changed
 - Fix flaky ConsulClusterStateChangesTest
@@ -19,6 +29,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [0.19.5]
 ### Changed 
 - Fixed issue with ConcurrentModificationException when using new ServicesState
+
+## [0.19.5]
+### Changed
+- Refactor reactor in RemoteServices 
 
 ## [0.19.3]
 ### Changed